It seems like a good idea to store the configuration files for all my dev tools online, so that's what I'm doing.
Almost a year ago now, I read a story on Hacker News about a guy that
was using Emacs as his window manager. That level of geekiness
intrigued me, so naturally I creeped his blog and github account. This
man, Howard Abrams, has an epic 'dot-files' repo. I thought it was
pretty cool at first, but I was too busy with other things at the time
to start a similar repo myself. Well, today for whatever reason I
remembered that repo, searched for it, reaffirmed its epicness, and
did a little more research on this dotfiles-in-git trend. After
reading a few blog posts on the benefits of keeping Emacs init files,
and other dot files in source control I was convinced. I still think
git + github are a bit overkill for the task, but its a lot easier for
me to use git + github instead of services like Google Drive or
Dropbox or even S3.